requirements.txt
ultralytics==8.3.3
pandas==2.2.2
pyarrow==15.0.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
minio==7.2.7
numpy==1.26.4
pyyaml==6.0.1
tqdm
open-clip-torch
scikit-learn==1.5.1
umap-learn==0.5.6